Let encode() in Wire trait return an error.
Given the way the crate Bytes works, all of the put_xx() methods
are infallible and the implementation of the encode() function of
the Wire trait may never fail, except for one case, if the size
of the encoded stream exceeds 16 bits which is the space we use
to encode the message length itself. Several options possible:
  1) use 32 or 64 octets and forget about the issue.
  2) let the encode() trait method return a Result, which will
     only be != Ok in one case.
  3) Define another trait method encode_fallible() and a default
     implementation that returns Ok, and implement it for RpcMsg
     only.

Opted for (2) in case we need to fail for other reasons in the
future.