Add method to limit number of objects in responses
Since we encode number of objects in a response with one octet,
we have to limit the number of objects we allow to be added.
An alternative would be that we allow any number of objects and
the encoding splits the response in multiple responses to
guarantee that the number in each does not exceed 255.