Add support for Get parameters
The wire format previously defined did not allow to specify what
type of information was to be retrieved. We may want to selectively
retrieve only certain type of objects. Also, within a type, we may
only want those that meet a certain criteria, such as referring to
an IP address, or MAC address, or, in case of routes, those of a
certain type or that use a particular next-hop, or encapsulation
(e.g. vni = 12345), or which belong to a certain vrf.

This patch adds a new Rust object GetFilter meant to be sent in
Get requests to specify the objects of interest to return.

The GetFilter is a collection of optional "match types" and their
possible values. Only 2 match types are implemented as a proof
of concept: match type Object type (to retrieve only objects of
a given type) and match type vrfId, to retrieve objects that are
scoped under a certain Vrf.

It also adds a new return value for a request 'ExpectMore' meant
to signal the requestor that more objects are to be expected.