Safer reads from buffers
The read methods of the bytes object (of the Buf trait) panic when
attempting to read more octets than those available. This is unde-
sirable since we would like to err instead. The prior code dealt
with this by explicitly calling a check_available() function.
However, that had to be called from many places and missing to do
so would create issues. Instead, define a new trait 'SafeReads'
which internally perform the check. This way, we make sure that
every read is checked and the code is much simpler to maintain.
The reader methods are exactly those of the Buf trait, preceded
by an "s".