clib: Add simple buffer implementation and utils.
A 'buffer_t' is the auxiliary data structure used to encode and
decode information in the wire format.

Since the wire protocol limits the max size of a message to 64KB,
the library or users could use a fixed-sized 64K buffer (e.g.
static or in stack). However, for flexibility, we allow buffer_t
to grow as needed. In order to limit heap re-allocations, a
buffer_t can be reutilized as many times as wanted, by calling
buffer_clear(). So, we get the best of two worlds.