Add small test tool cpmock
cpmock opens a unix socket and sends several types of requests to
to the echo (rust) server. For every message sent, it reads the
socket expecting to receive the same message it sent. This should
exercise the encoding / decoding of both the C and Rust
implementations and provide some coveragei to detect protocol
implementation divergence. The last message sent is a notification
to cause the other end (echo server) to terminate. Termination on
a notification is specific to the echo server, not necessarily the
protocol.