Add support for text fields (e.g. interface names)
Defines how variable-sized string/text fields are encoded in the
wire protocol. The encoding allows for strings of up to 255
characters, which should suffice in most cases and can otherwise be
easily changed.

This patch adds generic utils to encode / decode strings and
extends the interface object to include an interface name as an
example. In the particular case of kernel interface names, a limit
of 16 characters may exist. The wire format is not concerned nor
tries to enforce such a limit. This patch adds support for the
above ONLY in the Rust implementation.