Add sentinels to some enums and rename result ...
.. InvalidOperation to InvalidRequest.

Having sentinels is convenient in the C code. Manually add some
sentinels in proto.rs so that they get added to the C enums in
the cbindgen-generated C header clib/src/proto.h. Cbindgen can
automatically do that, but opted for manual addition since we may
not need/want all enums to have sentinels. Also, with the current
RenameRule in the EnumConfig of cbindgen that we use in build.rs,
the added sentinel is called "Sentinel" for all enums, which
causes duplicate definitions since we opted not to prefix enum
variants with the name of the enum.

Also renamed InvalidOperation to InvalidRequest as there may be
other reasons why a request is not accepted and the alternative of
enumerating explicitly all possibilities is not sensible at this
stage. Another option would be to completely remove this result
code and use just {Ok, Failure}. However, making the distinction
between Failure and Invalid may be helpful for diagnostics:

In general, RPC requests may fail in several ways. Without being
exhaustive, an rpc request may fail if:
  0) the reply never arrives. There's little we can do here, except
     for expiring the request after some time (which we have not
     defined in the wire protocol).
  1) the receiver is unable to decode it properly. This is a
     critical failure and, in general, it may not be wise/possible
     for the receiver to reply (hence the reason for defining
     "notifications" with which the receiver may provide details
     of the problem).
  2) the decoding of the request succeeded, but honoring it didn't
     for reasons not related to the syntax and conformance to the
     wire protocol. In other words, the receiver of the request did
     attempt to perform the operation, but it failed (e.g. due to
     not finding something, lack of resources, or some internal
     failure). In this case, the result code should be 'Failure',
     or other codes we may add if want to be more explicit.
  3) the decoding of the request succeeded but the receiver refused
     to or could not perform the operation because the request is
     invalid.