clib: add struct fmt_buff to stringify objects
struct fmt_buff is a simple buffer with a write offset that can
simplify stringifiers of objects as it tracks the buffer space
used against its capacity. The capacity is automatically adjusted
as data is written. A fmt_buff has to be initialized and finalized
since it allocates in heap. However, a single fmt_buff per thread
may be used as it can be cleared (rewinded) after / prior use.