Augment next-hop object with forwarding action.
We need to support blackhole routes and the current model implicit-
ly assumes that all routes have an action "forward". Change this
by adding a forward action field that defaults to "forward" but
can also be "drop". While for action "drop" no next-hop is actually
needed (and FRR will only allow for one), other actions might
include multiple next-hops. Therefore, the forwarding action is
added to the next-hop object instead of the route.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>