Add cache to rpc unix socket.
In the event of IO failures, we need to ensure that messages are
retransmitted, preserving their original order. It is not clear to
me if, in case async/Tokio was used for this, it would provide any
ordering guarantees. Crate futures seems to guarantee this with
struct 'FuturesUnordered', but we'll be using a dedicated thread
for IO in the dataplane. So...

..this patch adds object 'RpcCachedSock' which wraps a regular unix
socket (datagram) along with a message cache where messages that
cannot be sent are temporarily stored, preserving their order.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>