Extend ConnectInfo with synt
Synt (synchronization token) is a u64. The exact usage of this is
TBD, but the general intent is to use this field for dataplane to
detect if it restarted, so that the plugin needs not guess and
we retain the control of state refresh in dataplane, which allows
us to disable the behavior without needing to touch the plugin.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>